+++ src/main/java/com/github/mavogel/ilias/state/states/LoginState.java
@@ -25,19 +25,15 @@
  */
 package com.github.mavogel.ilias.state.states;
 
-import com.github.mavogel.client.ILIASSoapWebservicePortType;
 import com.github.mavogel.ilias.model.IliasAction;
 import com.github.mavogel.ilias.model.IliasNode;
 import com.github.mavogel.ilias.model.LoginConfiguration;
 import com.github.mavogel.ilias.state.ToolState;
 import com.github.mavogel.ilias.state.ToolStateMachine;
-import com.github.mavogel.ilias.utils.IliasUtils;
-import com.github.mavogel.ilias.wrapper.soap.SoapEndpoint;
+import com.github.mavogel.ilias.wrapper.AbstractIliasEndpoint;
+import com.github.mavogel.ilias.wrapper.EndpointBuilder;
 import org.apache.log4j.Logger;
 
-import javax.xml.rpc.ServiceException;
-import java.net.UnknownHostException;
-import java.rmi.RemoteException;
 import java.util.Collections;
 import java.util.List;
 
@@ -92,28 +88,11 @@ public class LoginState extends ToolState {
     protected void doExecute(final IliasAction nodesAndActions) {
         try {
             // choose SOAP or REST here
-            this.stateMachine.setIliasEndpoint(new SoapEndpoint(loginConfiguration));
+            AbstractIliasEndpoint iliasEndpoint = EndpointBuilder.build(EndpointBuilder.Type.SOAP, loginConfiguration);
+            this.stateMachine.setEndpoint(iliasEndpoint);
             LOG.info(String.format("Logged in successfully as '%s'", loginConfiguration.getUsername()));
         } catch (Exception e) {
             stateMachine.setState(stateMachine.getQuitState());
         }
-
-//        try {
-//            ILIASSoapWebservicePortType endpoint = IliasUtils.createWsEndpoint(loginConfiguration);
-//            stateMachine.setEndPoint(endpoint);
-//            stateMachine.setUserDataId(IliasUtils.getUserData(loginConfiguration, endpoint));
-//            LOG.info(String.format("Logged in successfully as '%s'", loginConfiguration.getUsername()));
-//        } catch (ServiceException e) {
-//            LOG.error("Could not establish webservice at '" + loginConfiguration.getEndpoint() + "'");
-//            stateMachine.setState(stateMachine.getQuitState());
-//        } catch (RemoteException e) {
-//            Throwable cause = e.getCause();
-//            if (cause instanceof UnknownHostException) {
-//                LOG.error("Could not establish connection to endpoint '" + loginConfiguration.getEndpoint()+ "'");
-//            } else {
-//                LOG.error("Error retrieving the user data: " + e.getMessage());
-//            }
-//            stateMachine.setState(stateMachine.getQuitState());
-//        }
     }
 }