+++ src/main/java/com/github/mavogel/ilias/state/states/QuitState.java
@@ -25,16 +25,13 @@
  */
 package com.github.mavogel.ilias.state.states;
 
-import com.github.mavogel.client.ILIASSoapWebservicePortType;
 import com.github.mavogel.ilias.model.IliasAction;
 import com.github.mavogel.ilias.model.IliasNode;
-import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.state.ToolState;
 import com.github.mavogel.ilias.state.ToolStateMachine;
 import com.github.mavogel.ilias.wrapper.IliasEndpoint;
 import org.apache.log4j.Logger;
 
-import java.rmi.RemoteException;
 import java.util.Collections;
 import java.util.List;
 
@@ -74,28 +71,12 @@ public class QuitState extends ToolState {
 
     @Override
     protected void doExecute(final IliasAction nodesAndActions) {
-        IliasEndpoint endpoint = this.stateMachine.getIliasEndpoint();
+        IliasEndpoint endpoint = this.stateMachine.getEndpoint();
         try {
             endpoint.logout();
         } catch (Exception e) {
             LOG.error("Could not log out: " + e.getMessage());
         }
         this.stateMachine.stop();
-
-        //        ILIASSoapWebservicePortType endPoint = this.stateMachine.getEndPoint();
-//        UserDataIds userDataIds = this.stateMachine.getUserDataIds();
-//        if (userDataIds != null && endPoint != null) {
-//            try {
-//                boolean isLoggedOut = endPoint.logout(userDataIds.getSid());
-//                if (isLoggedOut) {
-//                    LOG.info("Successfully logged out for sid: '" + userDataIds.getSid() + "'");
-//                } else {
-//                    LOG.error("Could not log out!");
-//                }
-//            } catch (RemoteException e) {
-//                LOG.error("Could not log out: " + e.getMessage());
-//            }
-//        }
-//        this.stateMachine.stop();
     }
 }