+++ src/main/java/com/github/mavogel/ilias/state/states/action/GrantFileUploadToGroupMembersAction.java
@@ -24,18 +24,13 @@ package com.github.mavogel.ilias.state.states.action;/*
  *  https://opensource.org/licenses/MIT
  */
 
-import com.github.mavogel.client.ILIASSoapWebservicePortType;
 import com.github.mavogel.ilias.model.IliasNode;
-import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.state.ChangeAction;
 import com.github.mavogel.ilias.state.ToolStateMachine;
 import com.github.mavogel.ilias.utils.IOUtils;
-import com.github.mavogel.ilias.utils.IliasUtils;
 import com.github.mavogel.ilias.wrapper.IliasEndpoint;
 import org.apache.log4j.Logger;
-import org.jdom.JDOMException;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
@@ -49,16 +44,13 @@ public class GrantFileUploadToGroupMembersAction implements ChangeAction {
     private static Logger LOG = Logger.getLogger(GrantFileUploadToGroupMembersAction.class);
 
     @Override
-    public void performAction(final IliasEndpoint iliasEndpoint,
+    public void performAction(final IliasEndpoint endpoint,
                               final Map<ToolStateMachine.ContextKey, List<IliasNode>> context,
                               final List<IliasNode> nodes) {
         LOG.info("Grant file upload to group members");
         if (confirm()) {
             try {
-                iliasEndpoint.grantFileUploadPermissionForMembers(nodes);
-//                IliasUtils.grantFileUploadPermissionForMembers(endpoint, sid, nodes);
-//            } catch (IOException | JDOMException e) {
-//                LOG.error("Error creating xml parser: " + e.getMessage());
+                endpoint.grantFileUploadPermissionForMembers(nodes);
             } catch (Exception e) {
                 LOG.error("Error granting file upload to group members: " + e.getMessage());
             }