+++ src/main/java/com/github/mavogel/ilias/utils/IOUtils.java
@@ -94,7 +94,6 @@ public class IOUtils {
             }
         }
 
-        scanner.close();
         return concatDigitsAndRanges(digits, ranges);
     }
 
@@ -224,7 +223,6 @@ public class IOUtils {
             }
         }
 
-        scanner.close();
         return userChoice;
     }
 
@@ -316,7 +314,6 @@ public class IOUtils {
             }
         }
 
-        scanner.close();
         return new RegistrationPeriod(registrationStart, registrationEnd);
     }
 
@@ -344,7 +341,6 @@ public class IOUtils {
             }
         }
 
-        scanner.close();
         return positiveInteger;
     }
 
@@ -372,7 +368,6 @@ public class IOUtils {
             }
         }
 
-        scanner.close();
         return choice;
     }
 
@@ -384,7 +379,6 @@ public class IOUtils {
     public static String readLine() {
         Scanner scanner = new Scanner(System.in);
         String nextLine = scanner.nextLine();
-        scanner.close();
         return StringUtils.deleteWhitespace(nextLine);
     }
 }