+++ src/test/java/com/github/mavogel/ilias/state/WorkflowTest.java
@@ -24,12 +24,14 @@ package com.github.mavogel.ilias.state;/*
  *  https://opensource.org/licenses/MIT
  */
 
-import com.github.mavogel.client.ILIASSoapWebserviceBindingStub;
 import com.github.mavogel.ilias.model.IliasNode;
 import com.github.mavogel.ilias.model.LoginConfiguration;
 import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.utils.IOUtils;
-import com.github.mavogel.ilias.utils.IliasUtils;
+import com.github.mavogel.ilias.wrapper.AbstractIliasEndpoint;
+import com.github.mavogel.ilias.wrapper.DisplayStatus;
+import com.github.mavogel.ilias.wrapper.EndpointBuilder;
+import com.github.mavogel.ilias.wrapper.soap.SoapEndpoint;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -38,23 +40,24 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import java.rmi.RemoteException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
 /**
  * Created by mavogel on 6/21/17.
- *
+ * <p>
  * Simulates complete workflows by mocking the ilias backend calls.
  */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({ToolStateMachine.class, WorkflowTest.class, IliasUtils.class, IOUtils.class})
+@PrepareForTest({ToolStateMachine.class, WorkflowTest.class, IOUtils.class,
+        AbstractIliasEndpoint.class, SoapEndpoint.class, EndpointBuilder.class})
 public class WorkflowTest {
 
     private LoginConfiguration loginConfiguration;
-    private ILIASSoapWebserviceBindingStub endpointMock;
+    private SoapEndpoint endpointMock;
     private UserDataIds userDateIds;
+//    private ToolStateMachine toolStateMachineMock;
 
     @Before
     public void setUp() throws Exception {
@@ -63,166 +66,140 @@ public class WorkflowTest {
 
     @Test
     public void shouldFindNoCoursesAndQuitThen() throws Exception {
-//        // == prepare
-//        final List<IliasNode> noCourses = Collections.emptyList();
-//
-//        // == train
-//        PowerMockito.when(IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN))
-//                .thenReturn(noCourses);
-//        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 2);
-//
-//        // == go
-//        createToolStateMachine().start();
-//
-//        // == verify
-//        PowerMockito.verifyStatic(Mockito.times(2));
-//        IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN);
-//        PowerMockito.verifyStatic(Mockito.times(2));
-//        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
+        // == prepare
+        final List<IliasNode> noCourses = Collections.emptyList();
+
+        // == train
+        PowerMockito.when(endpointMock.getCoursesForUser(DisplayStatus.ADMIN)).thenReturn(noCourses);
+        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 2);
+
+        // == go
+        createToolStateMachine().start();
+
+        // == verify
+        Mockito.verify(endpointMock, Mockito.times(2)).getCoursesForUser(DisplayStatus.ADMIN);
+        PowerMockito.verifyStatic(Mockito.times(2));
+        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
     }
 
-//    @Test
-//    public void shouldOneNCoursesAndQuitThen() throws Exception {
-//        // == prepare
-//        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
-//
-//        // == train
-//        PowerMockito.when(IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN))
-//                .thenReturn(courses);
-//        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 2);
-//
-//        // == go
-//        createToolStateMachine().start();
-//
-//        // == verify
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN);
-//        PowerMockito.verifyStatic(Mockito.times(2));
-//        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
-//    }
-//
-//    @Test
-//    public void shouldFind3GroupsInOneNCoursesAndRemoveUsers() throws Exception {
-//        // == prepare
-//        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
-//        final List<IliasNode> groups = Arrays.asList(
-//                new IliasNode(11, IliasNode.Type.GROUP, "Group 1"),
-//                new IliasNode(12, IliasNode.Type.GROUP, "Group 2"),
-//                new IliasNode(13, IliasNode.Type.GROUP, "Group 3"));
-//
-//        // == train
-//        PowerMockito.when(IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN))
-//                .thenReturn(courses);
-//        PowerMockito.when(IliasUtils.retrieveGroupRefIdsFromCourses(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), courses, 5)).thenReturn(groups);
-//        // 0: choose course; 1: actionsOnGroups; 1 quit
-//        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 1, 1);
-//
-//        PowerMockito.when(IOUtils.readAndParseChoicesFromUser(Mockito.anyList()))
-//                // 0,1,2: choosen groups;
-//                .thenReturn(Arrays.asList(0, 1, 2))
-//                // 1: remove users action
-//                .thenReturn(Arrays.asList(1));
-//
-//        PowerMockito.when(IOUtils.readAndParseUserConfirmation()).thenReturn(true);
-//        PowerMockito.doNothing().when(IliasUtils.class, "removeAllMembersFromGroups",
-//                endpointMock, userDateIds.getSid(), groups);
-//
-//        // == go
-//        createToolStateMachine().start();
-//
-//        // == verify
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN);
-//        PowerMockito.verifyStatic(Mockito.times(3));
-//        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
-//        PowerMockito.verifyStatic(Mockito.times(2));
-//        IOUtils.readAndParseChoicesFromUser(Mockito.anyList());
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.removeAllMembersFromGroups(endpointMock, userDateIds.getSid(), groups);
-//    }
-//
-//    @Test
-//    public void shouldFind4GroupsInOneNCoursesAndRemoveUsersAndUploadedMaterial() throws Exception {
-//        // == prepare
-//        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
-//        final List<IliasNode> groups = Arrays.asList(
-//                new IliasNode(11, IliasNode.Type.GROUP, "Group 1"),
-//                new IliasNode(12, IliasNode.Type.GROUP, "Group 2"),
-//                new IliasNode(13, IliasNode.Type.GROUP, "Group 3"),
-//                new IliasNode(14, IliasNode.Type.GROUP, "Group 4"));
-//        final List<IliasNode> files = Arrays.asList(
-//                new IliasNode(21, IliasNode.Type.FILE, "File 1"),
-//                new IliasNode(22, IliasNode.Type.FILE, "File 2"));
-//
-//        // == train
-//        PowerMockito.when(IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN))
-//                .thenReturn(courses);
-//        PowerMockito.when(IliasUtils.retrieveGroupRefIdsFromCourses(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), courses, 5)).thenReturn(groups);
-//        // 0: choose course; 1: actionsOnGroups; 1 quit
-//        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 1, 1);
-//
-//        PowerMockito.when(IOUtils.readAndParseChoicesFromUser(Mockito.anyList()))
-//                // 0,1,2,3: choosen groups;
-//                .thenReturn(Arrays.asList(0, 1, 2, 3))
-//                // 0: remove uploaded material; 1: remove users action
-//                .thenReturn(Arrays.asList(0, 1));
-//
-//        PowerMockito.when(IOUtils.readAndParseUserConfirmation()).thenReturn(true);
-//        PowerMockito.doNothing().when(IliasUtils.class, "removeAllMembersFromGroups",
-//                endpointMock, userDateIds.getSid(), groups);
-//        PowerMockito.when(IliasUtils.retrieveFileRefIdsFromGroups(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), groups)).thenReturn(files);
-//        PowerMockito.doNothing().when(IliasUtils.class, "deleteObjects",
-//                endpointMock, userDateIds.getSid(), files);
-//
-//        // == go
-//        createToolStateMachine().start();
-//
-//        // == verify
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.getCoursesForUser(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), IliasUtils.DisplayStatus.ADMIN);
-//
-//        PowerMockito.verifyStatic(Mockito.times(3));
-//        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
-//
-//        PowerMockito.verifyStatic(Mockito.times(2));
-//        IOUtils.readAndParseChoicesFromUser(Mockito.anyList());
-//
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.removeAllMembersFromGroups(endpointMock, userDateIds.getSid(), groups);
-//
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.retrieveFileRefIdsFromGroups(endpointMock, userDateIds.getSid(),
-//                userDateIds.getUserId(), groups);
-//
-//        PowerMockito.verifyStatic(Mockito.times(1));
-//        IliasUtils.deleteObjects(endpointMock, userDateIds.getSid(), files);
-//    }
+    @Test
+    public void shouldOneNCoursesAndQuitThen() throws Exception {
+        // == prepare
+        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
+
+        // == train
+        PowerMockito.when(endpointMock.getCoursesForUser(DisplayStatus.ADMIN)).thenReturn(courses);
+        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 2);
+
+        // == go
+        createToolStateMachine().start();
+
+        // == verify
+        Mockito.verify(endpointMock, Mockito.times(1)).getCoursesForUser(DisplayStatus.ADMIN);
+        PowerMockito.verifyStatic(Mockito.times(2));
+        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
+    }
+
+    @Test
+    public void shouldFind3GroupsInOneNCoursesAndRemoveUsers() throws Exception {
+        // == prepare
+        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
+        final List<IliasNode> groups = Arrays.asList(
+                new IliasNode(11, IliasNode.Type.GROUP, "Group 1"),
+                new IliasNode(12, IliasNode.Type.GROUP, "Group 2"),
+                new IliasNode(13, IliasNode.Type.GROUP, "Group 3"));
+
+        // == train
+        PowerMockito.when(endpointMock.getCoursesForUser(DisplayStatus.ADMIN)).thenReturn(courses);
+        PowerMockito.when(endpointMock.getGroupRefIdsFromCourses(courses.get(0))).thenReturn(groups);
+        // 0: choose course; 1: actionsOnGroups; 1 quit
+        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 1, 1);
+
+        PowerMockito.when(IOUtils.readAndParseChoicesFromUser(Mockito.anyList()))
+                // 0,1,2: choosen groups;
+                .thenReturn(Arrays.asList(0, 1, 2))
+                // 1: remove users action
+                .thenReturn(Arrays.asList(1));
+
+        PowerMockito.when(IOUtils.readAndParseUserConfirmation()).thenReturn(true);
+        PowerMockito.doNothing().when(endpointMock, "removeAllMembersFromGroups", groups);
+
+        // == go
+        createToolStateMachine().start();
+
+        // == verify
+        Mockito.verify(endpointMock, Mockito.times(1)).getCoursesForUser(DisplayStatus.ADMIN);
+        PowerMockito.verifyStatic(Mockito.times(3));
+        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
+        PowerMockito.verifyStatic(Mockito.times(2));
+        IOUtils.readAndParseChoicesFromUser(Mockito.anyList());
+        Mockito.verify(endpointMock, Mockito.times(1)).removeAllMembersFromGroups(groups);
+    }
+
+    @Test
+    public void shouldFind4GroupsInOneNCoursesAndRemoveUsersAndUploadedMaterial() throws Exception {
+        // == prepare
+        final List<IliasNode> courses = Arrays.asList(new IliasNode(1, IliasNode.Type.COURSE, "My Course"));
+        final List<IliasNode> groups = Arrays.asList(
+                new IliasNode(11, IliasNode.Type.GROUP, "Group 1"),
+                new IliasNode(12, IliasNode.Type.GROUP, "Group 2"),
+                new IliasNode(13, IliasNode.Type.GROUP, "Group 3"),
+                new IliasNode(14, IliasNode.Type.GROUP, "Group 4"));
+        final List<IliasNode> files = Arrays.asList(
+                new IliasNode(21, IliasNode.Type.FILE, "File 1"),
+                new IliasNode(22, IliasNode.Type.FILE, "File 2"));
+
+        // == train
+        PowerMockito.when(endpointMock.getCoursesForUser(DisplayStatus.ADMIN)).thenReturn(courses);
+        PowerMockito.when(endpointMock.getGroupRefIdsFromCourses(courses.get(0))).thenReturn(groups);
+        // 0: choose course; 1: actionsOnGroups; 1 quit
+        PowerMockito.when(IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList())).thenReturn(0, 1, 1);
+
+        PowerMockito.when(IOUtils.readAndParseChoicesFromUser(Mockito.anyList()))
+                // 0,1,2,3: choosen groups;
+                .thenReturn(Arrays.asList(0, 1, 2, 3))
+                // 0: remove uploaded material; 1: remove users action
+                .thenReturn(Arrays.asList(0, 1));
+
+        PowerMockito.when(IOUtils.readAndParseUserConfirmation()).thenReturn(true);
+        PowerMockito.doNothing().when(endpointMock, "removeAllMembersFromGroups", groups);
+        PowerMockito.when(endpointMock.getFilesFromGroups(groups)).thenReturn(files);
+        PowerMockito.doNothing().when(endpointMock, "deleteObjectNodes", files);
+
+        // == go
+        createToolStateMachine().start();
+
+        // == verify
+        Mockito.verify(endpointMock, Mockito.times(1)).getCoursesForUser(DisplayStatus.ADMIN);
+
+        PowerMockito.verifyStatic(Mockito.times(3));
+        IOUtils.readAndParseSingleChoiceFromUser(Mockito.anyList());
+
+        PowerMockito.verifyStatic(Mockito.times(2));
+        IOUtils.readAndParseChoicesFromUser(Mockito.anyList());
+
+        Mockito.verify(endpointMock, Mockito.times(1)).removeAllMembersFromGroups(groups);
+        Mockito.verify(endpointMock, Mockito.times(1)).getFilesFromGroups(groups);
+        Mockito.verify(endpointMock, Mockito.times(1)).deleteObjectNodes(files);
+    }
 
     ////////////////
     // HELPERS
     ////////////////
-    private void createEndpoint() throws javax.xml.rpc.ServiceException, RemoteException {
-        loginConfiguration = LoginConfiguration.asLDAPLogin("ep", "client", "user", "pwd", 5);
+    private void createEndpoint() throws Exception {
+        loginConfiguration = LoginConfiguration.asLDAPLogin("myEndpoint", "client_id", "user", "pwd", 5);
         userDateIds = new UserDataIds(123, "sid", loginConfiguration.getUsername());
-        endpointMock = PowerMockito.mock(ILIASSoapWebserviceBindingStub.class);
-        PowerMockito.mockStatic(IliasUtils.class);
         PowerMockito.mockStatic(IOUtils.class);
-        PowerMockito.when(IliasUtils.createWsEndpoint(loginConfiguration)).thenReturn(endpointMock);
-        PowerMockito.when(IliasUtils.getUserData(loginConfiguration, endpointMock)).thenReturn(userDateIds);
+        endpointMock = PowerMockito.mock(SoapEndpoint.class);
+        PowerMockito.mockStatic(EndpointBuilder.class);
+        PowerMockito.when(EndpointBuilder.build(EndpointBuilder.Type.SOAP, loginConfiguration)).thenReturn(endpointMock);
+
+//        toolStateMachineMock = PowerMockito.spy(new ToolStateMachine(loginConfiguration));
+//        PowerMockito.when(toolStateMachineMock.getUserDataIds()).thenReturn(userDateIds);
     }
 
     private ToolStateMachine createToolStateMachine() {
+//        return toolStateMachineMock;
         return new ToolStateMachine(loginConfiguration);
     }